fix: account issues
